create useLocale hook
