+++ src/components/DateTime/index.js
@@ -2,6 +2,9 @@ import React, { useState, useEffect, useRef } from 'react';
 import { useSelector } from 'react-redux';
 import { format, formatDistanceToNow, parseISO } from 'date-fns';
 
+import useDateFNSLocale from './useDateFNSLocale';
+import timeToString from './timeToString';
+
 /*
 Diplays a date & time in a local timezone.
 It also display the relative time in the hover/title.
@@ -30,39 +33,7 @@ The default format is `lll` -> `Aug 22, 2020 1:13 PM`
 */
 
 export function DateTime(props) {
-	const isMounted = useRef(null);
-	const [locale, setLocale] = useState('');
-	const language = useSelector(state => state.language.language)
-	
-	// Sets isMounted indicator to true
-	// when component has been mounted (componentDidMount analogue)
-	// and sets isMounted to false before component 
-	// is unmounted (componentWillUnmount analogue)
-	useEffect(() => {
-		isMounted.current = true;
-
-		return () => {
-			isMounted.current = false;
-		}
-	}, [])
-
-	useEffect(() => {
-		if (language && language !== "en") {
-			fetchLocale()
-		} else {
-			setLocale(undefined);
-		}
-	}, [language])
-
-	const fetchLocale = async () => {
-		// Don't remove comment inside import
-		// it's webpack's dynamic expression
-		const importedLocale = await import(
-			/* webpackMode: "lazy", webpackChunkName: "df-[index]" */
-			`date-fns/locale/${language}/index.js`
-		);
-		if (isMounted.current === true) setLocale(importedLocale.default);
-	}
+	const locale = useDateFNSLocale();
 
 	if ((props.value === null) || (props.value === undefined)) {
 		return (
@@ -79,9 +50,7 @@ export function DateTime(props) {
 		);
 	}
 	
-	const date = isNaN(props.value) ? format(parseISO(props.value), 'PPp', { locale: locale }) :
-		props.value > 9999999999 ? format(props.value, 'PPp', { locale: locale }) :
-		format(props.value * 1000, 'PPp', { locale: locale });
+	const date = timeToString(props.value, locale);
 
 	const dateFromNow = isNaN(props.value) ? formatDistanceToNow(parseISO(props.value), { locale: locale }) :
 		props.value > 9999999999 ? formatDistanceToNow(props.value, { locale: locale }) :
