+++ src/index.js
@@ -2,7 +2,6 @@ import "babel-polyfill";
 
 export { default as Application } from './containers/Application';
 
-export { DateTime } from './components/DateTime';
 export { Humanize } from './components/Humanize';
 export { DataTable } from './components/DataTable';
 export { ButtonWithAuthz } from './modules/auth/ButtonWithAuthz';
@@ -14,6 +13,8 @@ export { default as ControlledSwitch } from './components/ControlledSwitch';
 export { default as UncontrolledSwitch } from './components/UncontrolledSwitch';
 export { Credentials } from './components/Credentials';
 export { CellContentLoader, ChartLoader } from './components/ContentLoader';
-export { timeToString } from './components/timeToString';
+export { DateTime } from './components/DateTime';
+export { default as timeToString } from './components/DateTime/timeToString';
+export { default as useDateFNSLocale } from './components/DateTime/useDateFNSLocale';
 
 import "./style.scss";
